requirements.txt
streamlit~=1.36.0
tensorflow~=2.16.1
numpy~=1.26.4
requests~=2.32.3
keras~=3.3.3
pillow~=10.3.0